Share sum accumulation kernels with StandardSum via a seen out-param
