+++ vortex-array/src/aggregate_fn/fns/sum/primitive.rs
@@ -21,17 +21,24 @@ use crate::match_each_native_ptype;
 /// than 64 bits cannot overflow the 64-bit accumulator: `2^16 * (2^32 - 1) < 2^64`.
 const SUM_CHUNK: usize = 1 << 16;
 
-pub(super) fn accumulate_primitive(
+pub(crate) fn accumulate_primitive(
     inner: &mut SumState,
     p: &PrimitiveArray,
     ctx: &mut ExecutionCtx,
     skip_nans: bool,
+    seen: &mut bool,
 ) -> VortexResult<bool> {
     let mask = p.as_ref().validity()?.execute_mask(p.as_ref().len(), ctx)?;
     match mask.slices() {
         AllOr::None => Ok(false),
-        AllOr::All => accumulate_primitive_all(inner, p, skip_nans),
-        AllOr::Some(slices) => accumulate_primitive_valid(inner, p, slices, skip_nans),
+        AllOr::All => {
+            *seen |= !p.as_ref().is_empty();
+            accumulate_primitive_all(inner, p, skip_nans)
+        }
+        AllOr::Some(slices) => {
+            *seen |= !slices.is_empty();
+            accumulate_primitive_valid(inner, p, slices, skip_nans)
+        }
     }
 }
 
@@ -66,7 +73,7 @@ fn accumulate_primitive_all(
 /// Sum the values of a float slice into an `f64` accumulator. When `skip_nans` is set, NaN values
 /// are skipped to match the scalar `sum` semantics; otherwise any NaN poisons the accumulator to
 /// NaN. Floats cannot overflow the accumulator, so this never reports saturation.
-pub(super) fn sum_float_all<T: NativePType>(acc: &mut f64, slice: &[T], skip_nans: bool) {
+pub(crate) fn sum_float_all<T: NativePType>(acc: &mut f64, slice: &[T], skip_nans: bool) {
     if skip_nans {
         for &v in slice {
             if !v.is_nan() {
@@ -84,7 +91,7 @@ pub(super) fn sum_float_all<T: NativePType>(acc: &mut f64, slice: &[T], skip_nan
 /// chunks of [`SUM_CHUNK`] with a single checked add per chunk, which lets the inner loop vectorize
 /// to packed widening adds. `u64` input keeps a per-element checked add since a chunk of `u64`s
 /// could itself overflow. Returns `true` on overflow.
-pub(super) fn sum_unsigned_all<T>(acc: &mut u64, slice: &[T]) -> bool
+pub(crate) fn sum_unsigned_all<T>(acc: &mut u64, slice: &[T]) -> bool
 where
     T: NativePType + AsPrimitive<u64>,
 {
@@ -106,7 +113,7 @@ where
 }
 
 /// Signed counterpart of [`sum_unsigned_all`].
-pub(super) fn sum_signed_all<T>(acc: &mut i64, slice: &[T]) -> bool
+pub(crate) fn sum_signed_all<T>(acc: &mut i64, slice: &[T]) -> bool
 where
     T: NativePType + AsPrimitive<i64>,
 {
